+++ ext/opentelemetry-exporter-cloud-monitoring/tests/test_cloud_monitoring.py
@@ -17,26 +17,41 @@ from unittest import mock
 
 from google.api.label_pb2 import LabelDescriptor
 from google.api.metric_pb2 import MetricDescriptor
+from google.api.monitored_resource_pb2 import MonitoredResource
 from google.cloud.monitoring_v3.proto.metric_pb2 import TimeSeries
 
 from opentelemetry.exporter.cloud_monitoring import (
     MAX_BATCH_WRITE,
+    UNIQUE_IDENTIFIER_KEY,
     WRITE_INTERVAL,
     CloudMonitoringMetricsExporter,
 )
 from opentelemetry.sdk.metrics.export import MetricRecord
-from opentelemetry.sdk.metrics.export.aggregate import CounterAggregator
+from opentelemetry.sdk.metrics.export.aggregate import SumAggregator
+from opentelemetry.sdk.resources import Resource
 
 
 class UnsupportedAggregator:
     pass
 
 
+class MockMeter:
+    def __init__(self, resource=Resource.create_empty()):
+        self.resource = resource
+
+
 class MockMetric:
-    def __init__(self, name="name", description="description", value_type=int):
+    def __init__(
+        self,
+        name="name",
+        description="description",
+        value_type=int,
+        meter=None,
+    ):
         self.name = name
         self.description = description
         self.value_type = value_type
+        self.meter = meter or MockMeter()
 
 
 # pylint: disable=protected-access
@@ -114,7 +129,7 @@ class TestCloudMonitoringMetricsExporter(unittest.TestCase):
         )
 
         record = MetricRecord(
-            MockMetric(), (("label1", "value1"),), CounterAggregator(),
+            MockMetric(), (("label1", "value1"),), SumAggregator(),
         )
         metric_descriptor = exporter._get_metric_descriptor(record)
         client.create_metric_descriptor.assert_called_with(
@@ -149,7 +164,7 @@ class TestCloudMonitoringMetricsExporter(unittest.TestCase):
                     ("label3", 3),
                     ("label4", False),
                 ),
-                CounterAggregator(),
+                SumAggregator(),
             )
         )
         client.create_metric_descriptor.assert_called_with(
@@ -204,24 +219,41 @@ class TestCloudMonitoringMetricsExporter(unittest.TestCase):
             }
         )
 
-        counter_one = CounterAggregator()
-        counter_one.checkpoint = 1
-        counter_one.last_update_timestamp = (WRITE_INTERVAL + 1) * 1e9
+        resource = Resource(
+            labels={
+                "cloud.account.id": 123,
+                "host.id": "host",
+                "cloud.zone": "US",
+                "cloud.provider": "gcp",
+                "extra_info": "extra",
+                "gcp.resource_type": "gce_instance",
+                "not_gcp_resource": "value",
+            }
+        )
+
+        sum_agg_one = SumAggregator()
+        sum_agg_one.checkpoint = 1
+        sum_agg_one.last_update_timestamp = (WRITE_INTERVAL + 1) * 1e9
         exporter.export(
             [
                 MetricRecord(
-                    MockMetric(),
+                    MockMetric(meter=MockMeter(resource=resource)),
                     (("label1", "value1"), ("label2", 1),),
-                    counter_one,
+                    sum_agg_one,
                 ),
                 MetricRecord(
-                    MockMetric(),
+                    MockMetric(meter=MockMeter(resource=resource)),
                     (("label1", "value2"), ("label2", 2),),
-                    counter_one,
+                    sum_agg_one,
                 ),
             ]
         )
-        series1 = TimeSeries()
+        expected_resource = MonitoredResource(
+            type="gce_instance",
+            labels={"project_id": "123", "instance_id": "host", "zone": "US"},
+        )
+
+        series1 = TimeSeries(resource=expected_resource)
         series1.metric.type = "custom.googleapis.com/OpenTelemetry/name"
         series1.metric.labels["label1"] = "value1"
         series1.metric.labels["label2"] = "1"
@@ -230,7 +262,7 @@ class TestCloudMonitoringMetricsExporter(unittest.TestCase):
         point.interval.end_time.seconds = WRITE_INTERVAL + 1
         point.interval.end_time.nanos = 0
 
-        series2 = TimeSeries()
+        series2 = TimeSeries(resource=expected_resource)
         series2.metric.type = "custom.googleapis.com/OpenTelemetry/name"
         series2.metric.labels["label1"] = "value2"
         series2.metric.labels["label2"] = "2"
@@ -245,33 +277,33 @@ class TestCloudMonitoringMetricsExporter(unittest.TestCase):
         # Attempting to export too soon after another export with the exact
         # same labels leads to it being dropped
 
-        counter_two = CounterAggregator()
-        counter_two.checkpoint = 1
-        counter_two.last_update_timestamp = (WRITE_INTERVAL + 2) * 1e9
+        sum_agg_two = SumAggregator()
+        sum_agg_two.checkpoint = 1
+        sum_agg_two.last_update_timestamp = (WRITE_INTERVAL + 2) * 1e9
         exporter.export(
             [
                 MetricRecord(
                     MockMetric(),
                     (("label1", "value1"), ("label2", 1),),
-                    counter_two,
+                    sum_agg_two,
                 ),
                 MetricRecord(
                     MockMetric(),
                     (("label1", "value2"), ("label2", 2),),
-                    counter_two,
+                    sum_agg_two,
                 ),
             ]
         )
         self.assertEqual(client.create_time_series.call_count, 1)
 
         # But exporting with different labels is fine
-        counter_two.checkpoint = 2
+        sum_agg_two.checkpoint = 2
         exporter.export(
             [
                 MetricRecord(
                     MockMetric(),
                     (("label1", "changed_label"), ("label2", 2),),
-                    counter_two,
+                    sum_agg_two,
                 ),
             ]
         )
@@ -289,3 +321,116 @@ class TestCloudMonitoringMetricsExporter(unittest.TestCase):
                 mock.call(self.project_name, [series3]),
             ]
         )
+
+    def test_unique_identifier(self):
+        client = mock.Mock()
+        exporter1 = CloudMonitoringMetricsExporter(
+            project_id=self.project_id,
+            client=client,
+            add_unique_identifier=True,
+        )
+        exporter2 = CloudMonitoringMetricsExporter(
+            project_id=self.project_id,
+            client=client,
+            add_unique_identifier=True,
+        )
+        exporter1.project_name = self.project_name
+        exporter2.project_name = self.project_name
+
+        client.create_metric_descriptor.return_value = MetricDescriptor(
+            **{
+                "name": None,
+                "type": "custom.googleapis.com/OpenTelemetry/name",
+                "display_name": "name",
+                "description": "description",
+                "labels": [
+                    LabelDescriptor(
+                        key=UNIQUE_IDENTIFIER_KEY, value_type="STRING"
+                    ),
+                ],
+                "metric_kind": "GAUGE",
+                "value_type": "DOUBLE",
+            }
+        )
+
+        sum_agg_one = SumAggregator()
+        sum_agg_one.update(1)
+        metric_record = MetricRecord(MockMetric(), (), sum_agg_one,)
+        exporter1.export([metric_record])
+        exporter2.export([metric_record])
+
+        (
+            first_call,
+            second_call,
+        ) = client.create_metric_descriptor.call_args_list
+        self.assertEqual(first_call[0][1].labels[0].key, UNIQUE_IDENTIFIER_KEY)
+        self.assertEqual(
+            second_call[0][1].labels[0].key, UNIQUE_IDENTIFIER_KEY
+        )
+
+        first_call, second_call = client.create_time_series.call_args_list
+        self.assertNotEqual(
+            first_call[0][1][0].metric.labels[UNIQUE_IDENTIFIER_KEY],
+            second_call[0][1][0].metric.labels[UNIQUE_IDENTIFIER_KEY],
+        )
+
+    def test_extract_resources(self):
+        exporter = CloudMonitoringMetricsExporter(project_id=self.project_id)
+
+        self.assertIsNone(
+            exporter._get_monitored_resource(Resource.create_empty())
+        )
+        resource = Resource(
+            labels={
+                "cloud.account.id": 123,
+                "host.id": "host",
+                "cloud.zone": "US",
+                "cloud.provider": "gcp",
+                "extra_info": "extra",
+                "gcp.resource_type": "gce_instance",
+                "not_gcp_resource": "value",
+            }
+        )
+        expected_extract = MonitoredResource(
+            type="gce_instance",
+            labels={"project_id": "123", "instance_id": "host", "zone": "US"},
+        )
+        self.assertEqual(
+            exporter._get_monitored_resource(resource), expected_extract
+        )
+
+        resource = Resource(
+            labels={
+                "cloud.account.id": "123",
+                "host.id": "host",
+                "extra_info": "extra",
+                "not_gcp_resource": "value",
+                "gcp.resource_type": "gce_instance",
+                "cloud.provider": "gcp",
+            }
+        )
+        # Should throw when passed a malformed GCP resource dict
+        self.assertRaises(KeyError, exporter._get_monitored_resource, resource)
+
+        resource = Resource(
+            labels={
+                "cloud.account.id": "123",
+                "host.id": "host",
+                "extra_info": "extra",
+                "not_gcp_resource": "value",
+                "gcp.resource_type": "unsupported_gcp_resource",
+                "cloud.provider": "gcp",
+            }
+        )
+        self.assertIsNone(exporter._get_monitored_resource(resource))
+
+        resource = Resource(
+            labels={
+                "cloud.account.id": "123",
+                "host.id": "host",
+                "extra_info": "extra",
+                "not_gcp_resource": "value",
+                "cloud.provider": "aws",
+            }
+        )
+        self.assertIsNone(exporter._get_monitored_resource(resource))