+++ ext/opentelemetry-exporter-cloud-trace/src/opentelemetry/exporter/cloud_trace/__init__.py
@@ -44,6 +44,7 @@ import logging
 from typing import Any, Dict, List, Optional, Sequence, Tuple
 
 import google.auth
+import pkg_resources
 from google.cloud.trace_v2 import TraceServiceClient
 from google.cloud.trace_v2.proto.trace_pb2 import AttributeValue
 from google.cloud.trace_v2.proto.trace_pb2 import Span as ProtoSpan
@@ -51,9 +52,17 @@ from google.cloud.trace_v2.proto.trace_pb2 import TruncatableString
 from google.rpc.status_pb2 import Status
 
 import opentelemetry.trace as trace_api
+from opentelemetry.exporter.cloud_trace.version import (
+    __version__ as cloud_trace_version,
+)
+from opentelemetry.sdk.resources import Resource
 from opentelemetry.sdk.trace import Event
 from opentelemetry.sdk.trace.export import Span, SpanExporter, SpanExportResult
 from opentelemetry.sdk.util import BoundedDict
+from opentelemetry.trace.span import (
+    get_hexadecimal_span_id,
+    get_hexadecimal_trace_id,
+)
 from opentelemetry.util import types
 
 logger = logging.getLogger(__name__)
@@ -98,7 +107,6 @@ class CloudTraceSpanExporter(SpanExporter):
             # pylint: disable=broad-except
             except Exception as ex:
                 logger.error("Error when creating span %s", span, exc_info=ex)
-
         try:
             self.client.batch_write_spans(
                 "projects/{}".format(self.project_id), cloud_trace_spans,
@@ -123,15 +131,15 @@ class CloudTraceSpanExporter(SpanExporter):
 
         for span in spans:
             ctx = span.get_context()
-            trace_id = _get_hexadecimal_trace_id(ctx.trace_id)
-            span_id = _get_hexadecimal_span_id(ctx.span_id)
+            trace_id = get_hexadecimal_trace_id(ctx.trace_id)
+            span_id = get_hexadecimal_span_id(ctx.span_id)
             span_name = "projects/{}/traces/{}/spans/{}".format(
                 self.project_id, trace_id, span_id
             )
 
             parent_id = None
             if span.parent:
-                parent_id = _get_hexadecimal_span_id(span.parent.span_id)
+                parent_id = get_hexadecimal_span_id(span.parent.span_id)
 
             start_time = _get_time_from_ns(span.start_time)
             end_time = _get_time_from_ns(span.end_time)
@@ -141,6 +149,11 @@ class CloudTraceSpanExporter(SpanExporter):
                     "Span has more then %s attributes, some will be truncated",
                     MAX_SPAN_ATTRS,
                 )
+
+            # Span does not support a MonitoredResource object. We put the
+            # information into labels instead.
+            resources_and_attrs = _extract_resources(span.resource)
+            resources_and_attrs.update(span.attributes)
 
             cloud_trace_spans.append(
                 {
@@ -153,7 +166,9 @@ class CloudTraceSpanExporter(SpanExporter):
                     "end_time": end_time,
                     "parent_span_id": parent_id,
                     "attributes": _extract_attributes(
-                        span.attributes, MAX_SPAN_ATTRS
+                        resources_and_attrs,
+                        MAX_SPAN_ATTRS,
+                        add_agent_attr=True,
                     ),
                     "links": _extract_links(span.links),
                     "status": _extract_status(span.status),
@@ -167,14 +182,6 @@ class CloudTraceSpanExporter(SpanExporter):
 
     def shutdown(self):
         pass
-
-
-def _get_hexadecimal_trace_id(trace_id: int) -> str:
-    return "{:032x}".format(trace_id)
-
-
-def _get_hexadecimal_span_id(span_id: int) -> str:
-    return "{:016x}".format(span_id)
 
 
 def _get_time_from_ns(nanoseconds: int) -> Dict:
@@ -234,8 +241,8 @@ def _extract_links(links: Sequence[trace_api.Link]) -> ProtoSpan.Links:
                 "Link has more then %s attributes, some will be truncated",
                 MAX_LINK_ATTRS,
             )
-        trace_id = _get_hexadecimal_trace_id(link.context.trace_id)
-        span_id = _get_hexadecimal_span_id(link.context.span_id)
+        trace_id = get_hexadecimal_trace_id(link.context.trace_id)
+        span_id = get_hexadecimal_span_id(link.context.span_id)
         extracted_links.append(
             {
                 "trace_id": trace_id,
@@ -291,21 +298,64 @@ def _extract_events(events: Sequence[Event]) -> ProtoSpan.TimeEvents:
     )
 
 
+def _strip_characters(ot_version):
+    return "".join(filter(lambda x: x.isdigit() or x == ".", ot_version))
+
+
+OT_RESOURCE_LABEL_TO_GCP = {
+    "gce_instance": {
+        "cloud.account.id": "project_id",
+        "host.id": "instance_id",
+        "cloud.zone": "zone",
+    }
+}
+
+
+def _extract_resources(resource: Resource) -> Dict[str, str]:
+    if resource.labels.get("cloud.provider") != "gcp":
+        return {}
+    resource_type = resource.labels["gcp.resource_type"]
+    if resource_type not in OT_RESOURCE_LABEL_TO_GCP:
+        return {}
+    return {
+        "g.co/r/{}/{}".format(resource_type, gcp_resource_key,): str(
+            resource.labels[ot_resource_key]
+        )
+        for ot_resource_key, gcp_resource_key in OT_RESOURCE_LABEL_TO_GCP[
+            resource_type
+        ].items()
+    }
+
+
 def _extract_attributes(
-    attrs: types.Attributes, num_attrs_limit: int
+    attrs: types.Attributes,
+    num_attrs_limit: int,
+    add_agent_attr: bool = False,
 ) -> ProtoSpan.Attributes:
     """Convert span.attributes to dict."""
     attributes_dict = BoundedDict(num_attrs_limit)
-
+    invalid_value_dropped_count = 0
     for key, value in attrs.items():
         key = _truncate_str(key, 128)[0]
         value = _format_attribute_value(value)
 
-        if value is not None:
+        if value:
             attributes_dict[key] = value
+        else:
+            invalid_value_dropped_count += 1
+    if add_agent_attr:
+        attributes_dict["g.co/agent"] = _format_attribute_value(
+            "opentelemetry-python {}; google-cloud-trace-exporter {}".format(
+                _strip_characters(
+                    pkg_resources.get_distribution("opentelemetry-sdk").version
+                ),
+                _strip_characters(cloud_trace_version),
+            )
+        )
     return ProtoSpan.Attributes(
         attribute_map=attributes_dict,
-        dropped_attributes_count=len(attrs) - len(attributes_dict),
+        dropped_attributes_count=attributes_dict.dropped
+        + invalid_value_dropped_count,
     )