+++ opentelemetry-sdk/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+- Add support for resources and resource detector
+  ([#853](https://github.com/open-telemetry/opentelemetry-python/pull/853))
+
+## Version 0.10b0
+
+Released 2020-06-23
+
+- Rename CounterAggregator -> SumAggregator
+  ([#816](https://github.com/open-telemetry/opentelemetry-python/pull/816))
+
 ## 0.9b0
 
 Released 2020-06-10